+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# <a href='https://github.com/mrjackwills/flightbox_backend/releases/tag/v0.3.0'>v0.3.0</a>
+### 2025-08-18
+
+### Chores
++ dependencies updated, [3e300723](https://github.com/mrjackwills/flightbox_backend/commit/3e300723aeea7066300d9be582ae40f8d9b904cd)
+
+### Features
++ build for x86 musl, [87536dfc](https://github.com/mrjackwills/flightbox_backend/commit/87536dfca15126ec21e1ea2b3f3eedd059a2a212)
+
+### Refactors
++ screen controls removed, [21b58373](https://github.com/mrjackwills/flightbox_backend/commit/21b58373954043db23341fec156ecd0671d43969)
+
 # <a href='https://github.com/mrjackwills/flightbox_backend/releases/tag/v0.2.4'>v0.2.4</a>
 ### 2025-08-12